Move collapse to its own module
